OCE-64 - refactor + clean up some test.

+++ web/src/test/java/org/devgateway/ocds/web/rest/controller/AbstractEndPointControllerTest.java
@@ -6,14 +6,13 @@ import org.devgateway.ocds.persistence.mongo.repository.ReleaseRepository;
 import org.devgateway.ocds.persistence.mongo.spring.json.JsonImportPackage;
 import org.devgateway.ocds.persistence.mongo.spring.json.ReleasePackageJsonImport;
 import org.devgateway.toolkit.web.AbstractWebTest;
-import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.io.File;
-import java.util.Collection;
 import java.util.List;
 
 /**
@@ -25,35 +24,39 @@ import java.util.List;
 public abstract class AbstractEndPointControllerTest extends AbstractWebTest {
     protected static Logger logger = Logger.getLogger(AbstractEndPointControllerTest.class);
 
-    private Collection<Release> releases;
-
-    private List<Release> importedRelease;
-
     @Autowired
     private ReleaseRepository releaseRepository;
 
+    private static ReleaseRepository releaseRepositoryStatic;
+
     @Before
     public final void setUp() throws Exception {
+        // just run the setUp only once and be sure that the release collection is empty
+        if (testDataInitialized) {
+            return;
+        }
         releaseRepository.deleteAll();
+        releaseRepositoryStatic = releaseRepository;
 
         final ClassLoader classLoader = getClass().getClassLoader();
 
         final File file = new File(classLoader.getResource("json/endpoint-data-test.json").getFile());
         final JsonImportPackage releasePackageJsonImport = new ReleasePackageJsonImport(releaseRepository, file);
+        releasePackageJsonImport.importObjects();
 
-        this.releases = releasePackageJsonImport.importObjects();
-        this.importedRelease = releaseRepository.findAll();
+        testDataInitialized = true;
     }
 
-    @After
-    public final void tearDown() {
-        releaseRepository.deleteAll();
+    @AfterClass
+    public static final void tearDown() {
+        // be sure to clean up the release collection
+        releaseRepositoryStatic.deleteAll();
     }
 
     @Test
     public void testImportForEndpoints() {
         // just test that the import was done correctly
-        Assert.assertEquals(3, releases.size());
+        final List<Release> importedRelease = releaseRepository.findAll();
         Assert.assertNotNull(importedRelease);
         Assert.assertEquals(3, importedRelease.size());
     }
